pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "day08-langgraph-agent-lab"
version = "0.1.0"
description = "Starter skeleton for LangGraph agentic orchestration lab"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "rich>=13.7",
  "typer>=0.12",
  "langgraph>=0.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
]
sqlite = ["langgraph-checkpoint-sqlite>=2.0"]
postgres = ["langgraph-checkpoint-postgres>=2.0"]

[project.scripts]
agent-lab = "langgraph_agent_lab.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ANN"]
# ANN401: typing.Any is unavoidable when wrapping LangGraph's checkpointer/state objects.
ignore = ["ANN401"]

[tool.ruff.lint.per-file-ignores]
# Tests do not need full annotations; E402 is required because langgraph imports come after a
# pytest.mark.skipif guard.
"tests/*" = ["ANN001", "ANN201", "ANN202", "E402"]
# Scripts mutate sys.path before importing local packages, which triggers E402.
"scripts/*" = ["E402"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --strict-markers"
testpaths = ["tests"]
